Loading indicator during task loading
Resolves #29

Currently, the loading indicator is only shown during task loading, in future we might want to generalize it also for some other parts of our page (statistics etc.), but generally it's probably usually better to load page incrementally and the indicator only makes sense when the loading can take considerable time (and use it for small portions of the page, not full page). Task loading can potentially by quite long, but the main reason for using the full page loading indicator is the need of blockly workspace reset, which looks ugly, the loading indicator looks better, although there might be even better solution.